Fix build scripts

+++ build_scripts/04-install_sipsimple.sh
@@ -3,10 +3,18 @@
 version="5.2.3-mac"
 wget https://github.com/AGProjects/python3-sipsimple/archive/refs/tags/$version.tar.gz
 
-tar zxvf $version.tar.gz 
-rm $version.tar.gz 
+if [ -f $version.tar.gz ]; then
+    rm $version.tar.gz
+fi
+
+tar zxvf $version.tar.gz
+rm $version.tar.gz
 
 source activate_venv.sh
+
+if [ -d python3-sipsimple-$version ];then
+    rm python3-sipsimple-$version
+fi
 
 cd python3-sipsimple-$version
 
@@ -30,3 +38,4 @@ if [ $? -ne 0 ]; then
     cd
     exit 1
 fi
+
